fix boolean flag bug
when specifying a boolean flag (like `ha`), the corresponding parameter
was always passed as a command line argument, irrespective of the value
of the boolean.
It is likely that `'ha': False` (or similar) was not used often, since
simply omitting the key altogether would have the correct effect.
Still, a rather important bug!